reduce top hits aggregation memory consumption
move request structure out of top hits aggregation collector and use from the
passed structure instead

full
terms_many_with_top_hits    Memory: 58.2 MB (-43.64%)    Avg: 425.9680ms (-21.38%)    Median: 415.1097ms (-23.56%)    [395.5303ms .. 484.6325ms]
dense
terms_many_with_top_hits    Memory: 58.2 MB (-43.64%)    Avg: 440.0817ms (-19.68%)    Median: 432.2286ms (-21.10%)    [403.5632ms .. 497.7541ms]
sparse
terms_many_with_top_hits    Memory: 13.1 MB (-49.31%)    Avg: 33.3568ms (-32.19%)    Median: 33.0834ms (-31.86%)    [32.5126ms .. 35.7397ms]
multivalue
terms_many_with_top_hits    Memory: 58.2 MB (-43.64%)    Avg: 414.2340ms (-25.44%)    Median: 413.4144ms (-25.64%)    [403.9919ms .. 430.3170ms]

+++ src/aggregation/metric/extended_stats.rs
@@ -229,8 +229,7 @@ impl IntermediateExtendedStats {
                 * self.intermediate_stats.count as f64
                 * other.intermediate_stats.count as f64
                 / new_count as f64;
-        self.mean = (self.intermediate_stats.sum as f64 + other.intermediate_stats.sum as f64)
-            / new_count as f64;
+        self.mean = (self.intermediate_stats.sum + other.intermediate_stats.sum) / new_count as f64;
         self.sum_of_squares_elastic += other.sum_of_squares_elastic;
         self.delta_sum_for_squares_elastic += other.delta_sum_for_squares_elastic;
         self.intermediate_stats
